requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.3
apispec==6.3.0
apispec-webframeworks==0.5.2
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
boto3==1.28.31
botocore==1.31.31
charset-normalizer==3.2.0
click==8.1.7
flake8==6.1.0
flake8-import-order==0.18.2
Flask==2.3.2
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
flask-swagger-ui==4.11.1
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kafka-python==2.0.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mccabe==0.7.0
multidict==6.0.4
numpy==1.25.2
packaging==23.1
pika==1.3.2
pika-stubs==0.1.3
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
python-dateutil==2.8.2
PyYAML==6.0.1
s3transfer==0.6.2
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.7
yarl==1.9.2
